version 2.0.3
Bug Fix:

Terra is now working correctly
Fixed Incas improper spawns (mountain detection not working correctly)
Fixed Spawns too close to Tundra/Desert

New Features:
RTS mode is now also working on Continents and Continents-islands

Known issues:
On Terra or on Continents and Continents-island with RTS mode, you may end with a firaxis spawn
This is due to the fact that sometimes Firaxis does not correctly split the continents (they are linked by a thin band of ground).
I recommend to automatically remap in these conditions.
That issue will not be fixed as this is not the intended play (you want proper separations between the continents)